__error message__

+++ write_error.c
@@ -1,20 +1,19 @@
 #include "shell.h"
 
 /**
- * write_error - Write error message to stderr
- * @params: Parameters struct
- * @msg: Error message
+ * write_error - Write an error message to stderr.
+ * @params: Parameters struct containing program information.
+ * @msg: Error message to be written.
  *
- * Description: This function writes an error message to the standard error stream.
- * It takes a parameters struct and an error message as input. The function constructs
- * an error buffer and copies various strings into it. Finally, it writes the contents
- * of the error buffer to the standard error stream.
+ * This function writes an error message to the standard error stream (stderr).
+ * It constructs an error message including the program name, line number, and
+ * the provided error message, and then writes it to stderr.
  */
 void write_error(param_t *params, char *msg)
 {
     char errBuffer[1024] = {0};  /* Error buffer */
     char *writeHead = errBuffer;  /* Pointer to the current write position */
-    int len = 0;  /* Length of the error message */
+    size_t len = 0;  /* Length of the error message */
 
     len += string_copy(writeHead, params->argv[0]);  /* Copy program name */
     writeHead += len;
